finished option tests
