+++ src/test/java/com/mindlinksoft/recruitment/mychat/OptionsTests.java
@@ -51,80 +51,6 @@ public class OptionsTests {
         }
 
         /**
-         * Test filter by keyword option.
-         * 
-         * @throws FileNotFoundException Thrown when the the input is illegal
-         * @throws IOException           Thrown when the writting to the output file
-         *                               fails
-         */
-        @Test
-        public void testFilterByKeyword() throws FileNotFoundException, IOException {
-                ConversationExporter exporter = new ConversationExporter();
-
-                // fake configuration
-                ConversationExporterConfiguration configuration = new ConversationExporterConfiguration();
-                configuration.inputFilePath = "chat.txt";
-                configuration.outputFilePath = "testChat.json";
-                configuration.filterKeyword = "damn";
-
-                // fake conversation
-                Conversation conversation = generateFakeConversation();
-                // run through the rest of the conversation exporter
-                exporter.applyOptions(conversation, configuration);
-                exporter.writeConversation(conversation, configuration.outputFilePath);
-
-                Collection<Message> messages = conversation.getMessages();
-                Message[] messageArray = new Message[messages.size()];
-                messages.toArray(messageArray);
-
-                assertEquals(1, messageArray.length);
-
-                assertEquals(Instant.ofEpochSecond(1448470902), messageArray[0].getTimestamp());
-                assertEquals("Lokir", messageArray[0].senderId);
-                assertEquals("Damn you Stormcloaks. Skyrim was fine until you came along. Empire was nice and lazy. "
-                                + "If they hadn’t been looking for you, I could’ve stolen that horse and been half way "
-                                + "to Hammerfell. You there. You and me — we should be here. It’s these Stormcloaks the "
-                                + "Empire wants.", messageArray[0].getContent());
-        }
-
-        /**
-         * Test report generation.
-         * 
-         * @throws FileNotFoundException Thrown when the the input is illegal
-         * @throws IOException           Thrown when the writting to the output file
-         *                               fails
-         */
-        @Test
-        public void testReportGeneration() throws FileNotFoundException, IOException {
-                ConversationExporter exporter = new ConversationExporter();
-
-                // fake configuration
-                ConversationExporterConfiguration configuration = new ConversationExporterConfiguration();
-                configuration.inputFilePath = "chat.txt";
-                configuration.outputFilePath = "testChat.json";
-                configuration.report = true;
-
-                // fake conversation
-                Conversation conversation = generateFakeConversation();
-                // run through the rest of the conversation exporter
-                exporter.applyOptions(conversation, configuration);
-                exporter.writeConversation(conversation, configuration.outputFilePath);
-
-                List<User> report = conversation.getActivity();
-
-                assertEquals(3, report.size());
-
-                assertEquals("Ralof", report.get(0).getSender());
-                assertEquals(Integer.valueOf(2), report.get(0).getCount());
-
-                assertEquals("Imperial Soldier", report.get(1).getSender());
-                assertEquals(Integer.valueOf(1), report.get(1).getCount());
-
-                assertEquals("Lokir", report.get(2).getSender());
-                assertEquals(Integer.valueOf(1), report.get(2).getCount());
-        }
-
-        /**
          * Test report generation.
          * 
          * @throws FileNotFoundException Thrown when the the input is illegal
